+++ src/main/webapp/WEB-INF/views/home/main-new.html
@@ -1,18 +1,18 @@
 <!DOCTYPE html>
-<html lang="ko" xmlns:th="http://www.thymeleaf.org">
-
-<div th:replace="/common/layout/fragments/head-new :: head"></div>
-<link rel="stylesheet" th:href="@{/static/css/home.css}"/>
-
+<html lang="ko"
+      xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      layout:decorate="~{/common/layout/layout-new}">
+<!-- 메인 페이지 전용 스타일 -->
+<head layout:fragment="styles">
+    <link rel="stylesheet" th:href="@{/static/css/home.css}" />
+</head>
 <body>
-<div id="wrap">
-    <!-- 헤더 시작 -->
-    <div th:replace="/common/layout/fragments/header-new :: header"></div>
-    <!-- 헤더 끝 -->
-    
+<div layout:fragment="content">
+    <!-- 모달 -->
+    <th:block th:include="/home/fragment/findIdAndPwInclude :: myModal"></th:block>
     <!-- 메인 시작 -->
     <div id="container" class="mainContainer">
-        
         <!-- 스와이프 시작 -->
         <div class="mainCenterTop">
             <div class="swiper">
@@ -27,14 +27,11 @@
             </div>
         </div>
         <!--스와이프 끝-->
-        
         <!-- 메인 컨테이너 시작 -->
         <div class="containerWrap">
-            
             <!-- 좌측 영역 -->
             <div class="mainLeft">
                 <div class="mainLeftTop">
-                    
                     <!-- 로그인 된 상태 -->
                     <div th:if="${session.userData != null}">
                         <div class="profileWrap">
@@ -102,7 +99,7 @@
                             </div>
                         </div>
                     </div>
-                    
+
                     <!-- 로그인 전 상태 -->
                     <div th:if="${session.userData == null}">
                         <div class="loginWrap">
@@ -140,17 +137,13 @@
                             <button id="category-20" class="categoryBtn">미니룸</button>
                         </div>
                     </div>
-                    
                     <!-- 상품 리스트  -->
                     <div id="product-container" class="productContainer"></div>
-                
                 </div>
             </div>
             <!-- 좌측 영역 끝 -->
-            
             <!-- 중앙 영역 -->
             <div class="mainCenter">
-                
                 <!--급 상승 유저-->
                 <div class="topUserWrap">
                     <div class="topUserTitle">
@@ -343,7 +336,6 @@
                         </li>
                     </ul>
                 </div>
-                
                 <!-- 뉴스 영역 -->
                 <div class="newsWrap">
                     <p class="mainTitleText">실시간뉴스</p>
@@ -352,11 +344,9 @@
                 <!-- 뉴스 영역 끝-->
             </div>
             <!-- 중앙 영역 끝 -->
-            
             <!-- 오른쪽 영역 -->
             <div class="mainRight">
                 <div class="mainTopRight">
-                    
                     <!-- 날씨 영역 시작 -->
                     <div class="weather-container">
                         <div class="weather-content">
@@ -379,7 +369,6 @@
                     </div>
                     <!-- 날씨 영역 끝 -->
                 </div>
-                
                 <!-- 오늘의 운세 영역 시작 -->
                 <div class="todayLucyWrap">
                     <p class="mainTitleText">오늘의 운세 🍀</p>
@@ -388,29 +377,23 @@
                     </div>
                 </div>
                 <!-- 오늘의 운세 영역 끝 -->
-                
                 <!-- 영화 추천 영역 시작 -->
                 <div class="movieWrap">
                     <p class="mainTitleText">영화 추천</p>
                     <div id="movieWrapper" class="movieWrapper"></div>
                 </div>
                 <!-- 영화 추천 영역 끝 -->
-                
             </div>
             <!-- 오른쪽 영역 끝 -->
-            
         </div>
         <!-- 메인 컨테이너 끝 -->
-        
     </div>
     <!-- 메인 끝 -->
-    
-    <!-- 푸터 시작 -->
-    <div th:replace="/common/layout/fragments/footer-new :: footer"></div>
-    <!-- 푸터 끝 -->
+    <!--메인 끝-->
 </div>
-
-<div th:replace="/common/layout/fragments/script-new :: script"></div>
-<script th:type="module" th:src="@{/static/js/ptr/main-new.js}"></script>
+<!-- 메인 페이지 전용 JS -->
+<th:block layout:fragment="scripts">
+    <script th:type="module" th:src="@{/static/js/ptr/main-new.js}"></script>
+</th:block>
 </body>
 </html>
